Consolidate React contexts into one file
