+++ frontend/src/components/Contexts.js
@@ -2,5 +2,6 @@ import { createContext } from "react";
 
 const FileStructure = createContext();
 const ModalVariantsManager = createContext();
+const AdminContext = createContext();
 
-export { FileStructure, ModalVariantsManager };
+export { FileStructure, ModalVariantsManager, AdminContext };
